backend/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
langchain
langchain-community
langchain-ollama
langchain-google-genai
langgraph
pypdf
httpx
beautifulsoup4
google-search-results
python-dotenv
requests
chromadb
rank_bm25
pdfplumber
duckduckgo-search
google-search-results
